feat(*): 重构

+++ src/components/DropdownButton/index.tsx
@@ -1,7 +1,9 @@
 import {
   Button, Dropdown, Menu, Modal,
 } from 'antd';
-import { FunctionComponent, ReactElement, useState } from 'react';
+import {
+  FunctionComponent, ReactElement, useState,
+} from 'react';
 import {
   LockOutlined,
   PoweroffOutlined,
@@ -42,30 +44,40 @@ const DropdownButton: FunctionComponent<DropdownButtonProps> = ({
   };
 
   const handleLogout = () => {
-    localStorage.clear();
-    navigate('/user/login');
+    Modal.confirm({
+      title: '退出登录',
+      content: '确认要退出登录吗？退出登录后会将您名下的报告分配给其他医生。',
+      closable: true,
+      okText: '确定',
+      cancelText: '取消',
+      maskClosable: true,
+      onOk: () => {
+        localStorage.clear();
+        navigate('/user/login');
+      },
+    });
   };
 
   const menu = (
     <Menu style={{ width: 120 }}>
       <Menu.Item key="0" onClick={handleBackHome}>
         <HomeOutlined />
-        回到首页
+        <span style={{ marginLeft: '5px' }}>回到首页</span>
       </Menu.Item>
       <Menu.Divider />
       <Menu.Item key="1" onClick={handleToHistoryReport}>
         <FileOutlined />
-        历史报告
+        <span style={{ marginLeft: '5px' }}>历史报告</span>
       </Menu.Item>
       <Menu.Divider />
       <Menu.Item key="2" onClick={handleChangePassword}>
         <LockOutlined />
-        修改密码
+        <span style={{ marginLeft: '5px' }}>修改密码</span>
       </Menu.Item>
       <Menu.Divider />
       <Menu.Item key="3" onClick={handleLogout}>
         <PoweroffOutlined />
-        退出登录
+        <span style={{ marginLeft: '5px' }}>退出登录</span>
       </Menu.Item>
     </Menu>
   );
